Revert "rossum-agent: Put tools into a sole module & improve test coverage"
