+++ rossum-agent/rossum_agent/__init__.py
@@ -4,6 +4,6 @@ from __future__ import annotations
 
 from rossum_agent.agent import AgentConfig, AgentStep, RossumAgent, create_agent
 
-__version__ = "0.3.0dev"
+__version__ = "0.2.7dev"
 
 __all__ = ["AgentConfig", "AgentStep", "RossumAgent", "create_agent"]
