--- rossum-agent/tests/tools/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the rossum_agent.tools package."""
